More helper file updates
